requirements.txt
beatserver==0.0.7
aioredis==1.3.1
channels==2.4.0
channels-redis==2.4.2
daphne==2.4.1
django==3.0.6
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
drf-yasg==1.17.1
pytest==6.0.1
pytest-asyncio==0.14.0
pytest-django==3.9.0
uwsgi
mysqlclient
